pkg/endpoint: move state-related code into separate file
The file endpoint.go is massive. Put endpoint-state-related code into its own file.

Signed-off by: Ian Vernon <ian@cilium.io>
